server/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.0.5
djangorestframework==3.12.4
idna==2.10
jwt==1.2.0
psycopg2==2.8.6
pycparser==2.20
python-environ==0.4.54
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.4
